settlementservice/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
PyMySQL
python-dotenv
requests
